Compiler warning fixed
Compiler warning 'return value may not fit into declared variable' fixed

+++ common/client_socket.cpp
@@ -63,7 +63,7 @@ ClientSocket& ClientSocket::operator=(ClientSocket&& other) {
 
 core::optional<size_t> ClientSocket::send(const void* payload, size_t size) {
     if (!isValid()) return false;
-    int res;
+    ssize_t res;
     if ((res = ::send(mSocketDesc, payload, size, 0)) < 0) {
         perror("Could not sent");
         return core::nullopt;
